% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ions_minimal.R
\name{combine_ions_minimal}
\alias{combine_ions_minimal}
\title{Combine ions into minimal AND/OR expression with sorting}
\usage{
combine_ions_minimal(ion_lists)
}
\arguments{
\item{ion_lists}{List of character vectors representing ion combinations}
}
\value{
Single character string representing nested AND/OR expression
}
\description{
Combine ions into minimal AND/OR expression with sorting
}
