% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_list_of_queries.R
\name{contains_all_mz}
\alias{contains_all_mz}
\title{Check if spectrum contains all target m/z values}
\usage{
contains_all_mz(spec_mz, target_mz, dalton, ppm)
}
\arguments{
\item{spec_mz}{Numeric vector of spectrum m/z values}

\item{target_mz}{Numeric vector of target m/z values to find}

\item{dalton}{Tolerance in Dalton}

\item{ppm}{Tolerance in parts per million}
}
\value{
Logical indicating if all targets are found
}
\description{
Check if spectrum contains all target m/z values
}
\examples{
NULL
}
