% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create_matrix}
\alias{create_matrix}
\title{Create matrix}
\usage{
create_matrix(spectra, zero_val = 0, name)
}
\arguments{
\item{spectra}{Spectra}

\item{zero_val}{Zero value}

\item{name}{Name}
}
\description{
Create matrix
}
\examples{
NULL
}
