% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matrix.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Filter matrix}
\usage{
filter_matrix(matrix, n)
}
\arguments{
\item{matrix}{Matrix to filter}

\item{n}{Minimum number of non-zero values per column}
}
\value{
Filtered matrix with columns having at least n non-zero values
}
\description{
Filter matrix
}
\examples{
NULL
}
