% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_binned_mzs.R
\name{fix_binned_mzs}
\alias{fix_binned_mzs}
\title{Fix binned mzs}
\usage{
fix_binned_mzs(binned_m, original_mzs, dalton, ppm, decimals)
}
\arguments{
\item{binned_m}{Binned matrix}

\item{original_mzs}{Original m/z values from spectra}

\item{dalton}{Tolerance in Dalton}

\item{ppm}{Tolerance in parts per million}

\item{decimals}{Number of decimal places for rounding}
}
\value{
Matrix with corrected m/z column names
}
\description{
Fix binned mzs
}
\examples{
NULL
}
