% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_combinations.R
\name{generate_combinations}
\alias{generate_combinations}
\title{Generate combinations}
\usage{
generate_combinations(x, max_ions)
}
\arguments{
\item{x}{Vector of elements to combine}

\item{max_ions}{Maximum number of ions in combinations}
}
\value{
List of all combinations
}
\description{
Generate combinations
}
\examples{
NULL
}
