% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_mzs.R
\name{harmonize_mzs}
\alias{harmonize_mzs}
\title{Harmonize mzs}
\usage{
harmonize_mzs(spectra, dalton, ppm)
}
\arguments{
\item{spectra}{Spectra object}

\item{dalton}{Tolerance in Dalton}

\item{ppm}{Tolerance in parts per million}
}
\value{
Spectra with harmonized m/z values
}
\description{
Harmonize mzs
}
\examples{
NULL
}
