% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yaml.R
\name{parse_yaml}
\alias{parse_yaml}
\title{Parse YAML}
\usage{
parse_yaml(file)
}
\arguments{
\item{file}{The path to the YAML file}
}
\value{
A list containing the parameters specified in the YAML file
}
\description{
This function parses YAML files
}
\examples{
NULL
}
