% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_list_of_queries.R
\name{perform_list_of_queries}
\alias{perform_list_of_queries}
\title{Perform list of queries}
\usage{
perform_list_of_queries(index, ions_list, query_data, dalton, ppm)
}
\arguments{
\item{index}{Index of the ion list to process}

\item{ions_list}{List of ions for queries}

\item{query_data}{Data frame with columns \code{mz}, \code{precursor_mz}, and \code{skeleton}}

\item{dalton}{Tolerance in Dalton}

\item{ppm}{Tolerance in parts per million}
}
\value{
Data frame with target and value columns
}
\description{
Perform list of queries
}
\examples{
NULL
}
