% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_list_of_queries.R
\name{perform_query}
\alias{perform_query}
\title{Perform query}
\usage{
perform_query(all_mz, precursor_mz, frags, nls, dalton, ppm)
}
\arguments{
\item{all_mz}{List of numeric vectors of m/z values for each spectrum}

\item{precursor_mz}{Numeric vector of precursor m/z values for each spectrum}

\item{frags}{Fragment masses to search for}

\item{nls}{Neutral losses to search for}

\item{dalton}{Tolerance in Dalton}

\item{ppm}{Tolerance in parts per million}
}
\value{
Filtered spectra object
}
\description{
Perform query
}
\examples{
NULL
}
