% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_to_queries.R
\name{spectra_to_queries}
\alias{spectra_to_queries}
\title{Spectra to queries}
\usage{
spectra_to_queries(
  spectra = NULL,
  export = "data/interim/queries.tsv",
  dalton = 0.01,
  decimals = 4L,
  intensity_min = 0,
  ions_max = 10L,
  n_skel_min = 5L,
  n_spec_min = 3L,
  ppm = 30,
  mcc_min = 0,
  precision_min = 0,
  recall_min = 0,
  zero_val = 0
)
}
\arguments{
\item{spectra}{Spectra object or path to MGF file. If \code{NULL}, loads the example
\code{spectra} dataset. If the string \code{"grouped"}, loads the grouped example
spectra. Otherwise, a path to an MGF file or a Spectra object.}

\item{export}{Export path}

\item{dalton}{Tolerance in Dalton. Default to 0.01}

\item{decimals}{Number of decimals for rounding. Default to 4L}

\item{intensity_min}{Minimal intensity. Default to 0.0}

\item{ions_max}{Maximal number of ions in the query. Default to 10L}

\item{n_skel_min}{Minimal number of individuals per skeleton. Default to 5L}

\item{n_spec_min}{Minimal number of individuals where a signal has to be found. Default to 3L}

\item{ppm}{Tolerance in parts per million Default to 30.0}

\item{mcc_min}{Minimal single-ion Matthews Correlation Coefficient. Default to 0.0}

\item{precision_min}{Minimal single ion precision (pre-filter before MCC). Default to 0.0}

\item{recall_min}{Minimal single ion recall (pre-filter before MCC). Default to 0.0}

\item{zero_val}{Zero value for intensity. Default to 0.0}
}
\value{
A file with diagnostic query ions
}
\description{
This function converts spectra to diagnostic ion queries.
}
\examples{
\dontrun{
  # Run with default spectra
  spectra_to_queries()

  # Or explicitly load and pass spectra_grouped
  data(spectra_grouped)
  spectra_to_queries(spectra = spectra_grouped)
}
}
