% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraToQueries-package.R
\docType{package}
\name{SpectraToQueries-package}
\alias{SpectraToQueries}
\alias{SpectraToQueries-package}
\title{SpectraToQueries: Spectra to queries}
\description{
SpectraToQueries provides the infrastructure to translate spectra to queries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/spectra-to-knowledge/spectra-to-queries}
  \item Report bugs at \url{https://github.com/spectra-to-knowledge/spectra-to-queries/issues}
}

}
\author{
\strong{Maintainer}: Adriano Rutz \email{rutz@imsb.biol.ethz.ch} (\href{https://orcid.org/0000-0003-0443-9902}{ORCID})

}
\keyword{internal}
