% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create_matrix}
\alias{create_matrix}
\title{Title}
\usage{
create_matrix(spectra, zero_val = ZERO_VAL, name)
}
\arguments{
\item{spectra}{Spectra}

\item{zero_val}{Zero value}

\item{name}{Name}
}
\description{
Title
}
\examples{
NULL
}
