% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_binned_mzs.R
\name{fix_binned_mzs}
\alias{fix_binned_mzs}
\title{Fix binned mzs}
\usage{
fix_binned_mzs(binned_m, original_mzs, dalton, decimals)
}
\arguments{
\item{binned_m}{binned_matrix}

\item{original_mzs}{original mzs}

\item{dalton}{dalton}

\item{decimals}{decimals}
}
\description{
Fix binned mzs
}
\examples{
NULL
}
