% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_combinations.R
\name{generate_combinations}
\alias{generate_combinations}
\title{Generate combinations}
\usage{
generate_combinations(x, max_ions)
}
\arguments{
\item{x}{x}
}
\description{
Generate combinations
}
\examples{
NULL
}
