% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_list_of_queries.R
\name{perform_list_of_queries}
\alias{perform_list_of_queries}
\title{Perform list of queries}
\usage{
perform_list_of_queries(index, ions_list, spectra)
}
\arguments{
\item{index}{Index}

\item{ions_list}{Ions list}

\item{spectra}{Spectra}
}
\description{
Perform list of queries
}
\examples{
NULL
}
