% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_query.R
\name{perform_query}
\alias{perform_query}
\title{Perform query}
\usage{
perform_query(spectra, frags, nls, dalton, ppm)
}
\arguments{
\item{spectra}{Spectra object}

\item{frags}{Fragment masses to search for}

\item{nls}{Neutral losses to search for}

\item{dalton}{Tolerance in Dalton}

\item{ppm}{Tolerance in parts per million}
}
\value{
Filtered spectra object
}
\description{
Perform query
}
\examples{
NULL
}
