% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_query.R
\name{perform_query}
\alias{perform_query}
\title{Perform query}
\usage{
perform_query(spectra, frags, nls, dalton, ppm)
}
\arguments{
\item{spectra}{Spectra}

\item{frags}{Fragments}

\item{nls}{Neutral losses}

\item{dalton}{Dalton}

\item{ppm}{PPM}
}
\description{
Perform query
}
\examples{
NULL
}
